1. Add base support 3dfx Glide API 2. Refactoring (move API to APIs folder) 3. Fix WaitEvent (Thanks Bobbie)

+++ Examples/OpenGL_SuperBible/Chapter_2_GLRect.cpp
@@ -11,7 +11,7 @@
 #include <LDL/Core/NumberToString.hpp>
 #include <LDL/Graphics/Window.hpp>
 #include <LDL/Graphics/Render.hpp>
-#include <LDL/OpenGL/OpenGL1_2.hpp>
+#include <LDL/APIs/OpenGL/OpenGL1_2.hpp>
 
 using namespace LDL::Graphics;
 using namespace LDL::Events;
@@ -102,7 +102,7 @@ int main()
 
 			render.Begin();
 
-			ChangeSize(window.Size().x, window.Size().y);
+			ChangeSize((int)window.Size().x, (int)window.Size().y);
 
 			RenderScene();
